+++ modules/gov/handle_block.go
@@ -8,8 +8,6 @@ import (
 	tmctypes "github.com/tendermint/tendermint/rpc/core/types"
 
 	"github.com/rs/zerolog/log"
-
-	"github.com/forbole/bdjuno/v2/types"
 )
 
 // HandleBlock implements modules.BlockModule
@@ -30,47 +28,6 @@ func (m *Module) HandleBlock(
 	return nil
 }
 
-// updateParams updates the governance parameters for the given height
-func (m *Module) updateParams(height int64) error {
-	depositParams, err := m.source.DepositParams(height)
-	if err != nil {
-		return fmt.Errorf("error while getting gov deposit params: %s", err)
-	}
-
-	votingParams, err := m.source.VotingParams(height)
-	if err != nil {
-		return fmt.Errorf("error while getting gov voting params: %s", err)
-	}
-
-	tallyParams, err := m.source.TallyParams(height)
-	if err != nil {
-		return fmt.Errorf("error while getting gov tally params: %s", err)
-	}
-
-	return m.db.SaveGovParams(types.NewGovParams(
-		types.NewVotingParams(votingParams),
-		types.NewDepositParam(depositParams),
-		types.NewTallyParams(tallyParams),
-		height,
-	))
-}
-
-// updateProposals updates the proposals
-func (m *Module) updateProposals(height int64, blockVals *tmctypes.ResultValidators) error {
-	ids, err := m.db.GetOpenProposalsIds()
-	if err != nil {
-		log.Error().Err(err).Str("module", "gov").Msg("error while getting open ids")
-	}
-
-	for _, id := range ids {
-		err = m.UpdateProposal(height, blockVals, id)
-		if err != nil {
-			return fmt.Errorf("error while updating proposal: %s", err)
-		}
-	}
-	return nil
-}
-
 // updateParamChangeProposals updates the params if a ParamChangeProposal passed
 func (m *Module) updateParamChangeProposals(height int64) error {
 	// Get the parameter change proposals
@@ -86,5 +43,21 @@ func (m *Module) updateParamChangeProposals(height int64) error {
 		}
 	}
 
+	return nil
+}
+
+// updateProposals updates the proposals
+func (m *Module) updateProposals(height int64, blockVals *tmctypes.ResultValidators) error {
+	ids, err := m.db.GetOpenProposalsIds()
+	if err != nil {
+		log.Error().Err(err).Str("module", "gov").Msg("error while getting open ids")
+	}
+
+	for _, id := range ids {
+		err = m.UpdateProposal(height, blockVals, id)
+		if err != nil {
+			return fmt.Errorf("error while updating proposal: %s", err)
+		}
+	}
 	return nil
 }
